requirements.txt
flask
pillow
gunicorn
opencv-python-headless
numpy